+++ packages/shared-fixtures/fixtures/typescript/basics/class-with-constructor-and-parameter-property-with-modifiers.src.ts
@@ -1,3 +1,0 @@
-class SpecializedComponent extends SomeComponent {
-  constructor(protected override readonly param: string) {}
-}