+++ packages/shared-fixtures/fixtures/typescript/basics/type-assertion-with-guard-in-interface.src.ts
@@ -1,3 +1,0 @@
-interface AssertFoo {
-  isString(node: any): asserts node is string;
-}