+++ packages/shared-fixtures/fixtures/typescript/errorRecovery/interface-method-export.src.ts
@@ -1,4 +1,0 @@
-interface Foo {
-    export g(bar: string): void;
-}
-