+++ packages/shared-fixtures/fixtures/typescript/types/function-with-rest.src.ts
@@ -1,1 +1,0 @@
-let f: (...a: number[]) => void;