+++ packages/utils/src/ts-eslint/SourceCode.ts
@@ -389,10 +389,25 @@ namespace SourceCode {
   }
 
   export type FilterPredicate = (token: TSESTree.Token) => boolean;
+  export type GetFilterPredicate<TFilter, TDefault> =
+    // https://github.com/prettier/prettier/issues/14275
+    // prettier-ignore
+    TFilter extends ((
+      token: TSESTree.Token,
+    ) => token is infer U extends TSESTree.Token)
+      ? U
+      : TDefault;
+  export type GetFilterPredicateFromOptions<TOptions, TDefault> =
+    TOptions extends { filter?: FilterPredicate }
+      ? GetFilterPredicate<TOptions['filter'], TDefault>
+      : GetFilterPredicate<TOptions, TDefault>;
 
   export type ReturnTypeFromOptions<T> = T extends { includeComments: true }
-    ? TSESTree.Token
-    : Exclude<TSESTree.Token, TSESTree.Comment>;
+    ? GetFilterPredicateFromOptions<T, TSESTree.Token>
+    : GetFilterPredicateFromOptions<
+        T,
+        Exclude<TSESTree.Token, TSESTree.Comment>
+      >;
 
   export type CursorWithSkipOptions =
     | number